requirements.txt
librosa==0.9.2
lightning==2.0.0
lightning_fabric==2.0.0
matplotlib==3.5.3
numpy==1.23.5
scipy==1.9.1
soundfile==0.11.0
torchmetrics==0.11.4
tqdm==4.65.0
transformers==4.27.4
